+++ code/master/api/tethys_api/views.py
@@ -8,7 +8,7 @@ from rest_framework.response import Response
 from .models import *
 from .serializers import *
 from .common import *
-from .globals import LAST_DATA_UPDATE as lastUpdateTimestamp
+from .globals import setLastDataUpdate, getLastDataUpdate
 
 # Get the absolute path of the core directory
 root_path = os.path.abspath(os.path.join(os.path.dirname(__file__), '../..'))
@@ -23,8 +23,7 @@ import globals.config as tethysConfig
 
 # :::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::
 def setLastDataUpdateNow():
-    global lastUpdateTimestamp
-    lastUpdateTimestamp = datetime.now()
+    setLastDataUpdate()
 
 
 # :::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::
@@ -47,8 +46,7 @@ def lastUpdate(request):
 
     if request.method == 'GET':
 
-        global lastUpdateTimestamp
-        return Response({'timestamp': lastUpdateTimestamp})
+        return Response({'timestamp': getLastDataUpdate()})
 
 
 # :::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::